Rename satellite modules to sgp4.py and kepler.py
Remove orbit predictor Location class as inherited base class
